LPS-89326 Move PermissionPropagator and its base class. Maintain backwards compatibility.
