+++ portal-kernel/src/com/liferay/portal/kernel/model/Portlet.java
@@ -632,7 +632,7 @@ public interface Portlet extends PortletModel, PersistedModel {
 	*
 	* @return the permission propagator instance of the portlet
 	*/
-	public com.liferay.portal.kernel.security.permission.PermissionPropagator getPermissionPropagatorInstance();
+	public com.liferay.portal.kernel.security.permission.propagator.PermissionPropagator getPermissionPropagatorInstance();
 
 	/**
 	* Returns the plugin ID of the portlet.
